# api/openapi.yaml
openapi: 3.0.3
info:
  title: IMS ERP System API
  description: |
    Complete ERP system API with microservices architecture
    
    ## Architecture
    - **API Gateway** (8080) - Central entry point for all client requests
    - **Auth Service** (8081) - Authentication and authorization
    - **Client Services** (8082) - Client management (CQRS pattern)
    - **Invoice Service** (8083) - Invoice management
    - **Payment Service** (8084) - Payment processing
    - **Product Service** (8085) - Product catalog
    - **Order Service** (8086) - Order management
    - **Inventory Service** (8087) - Inventory control
    
    ## Authentication
    All protected endpoints require Bearer token authentication.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to 1000 requests per minute per tenant.
    
    ## Error Handling
    All errors follow standard HTTP status codes with detailed error messages.
  version: 1.0.0
  contact:
    name: IMS ERP Support
    email: support@ims-erp.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.ims-erp.example.com/api/v1
    description: Production server
...
paths:
...
  /invoices:
    get:
      tags:
        - Invoices
      summary: List invoices
      parameters:
        - in: query
          name: tenantId
          required: true
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, pending, sent, paid, overdue, cancelled, refunded]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of invoices
    post:
      tags:
        - Invoices
      summary: Create invoice
      parameters:
        - in: query
          name: tenantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Invoice created successfully
...
components:
  schemas:
...
    CreateInvoiceRequest:
      type: object
      required:
        - clientId
        - currency
        - paymentTerm
      properties:
        clientId:
          type: string
          format: uuid
        currency:
          type: string
        paymentTerm:
          type: string
          enum: [due_on_receipt, net_15, net_30, net_45, net_60, end_of_month]
        issueDate:
          type: string
          format: date
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineRequest'
        notes:
          type: string
        terms:
          type: string
    InvoiceLineRequest:
      type: object
      required:
        - description
        - quantity
        - unitPrice
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        discount:
          type: number
          default: 0
        taxRate:
          type: number
          default: 0
        productId:
          type: string
          format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []